pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = "Backend para ingesta y búsqueda de licitaciones de Mercado Público Chile"
authors = ["Antigravity <antigravity@example.com>"]
readme = "README.md"
packages = [
    {include = "app"}
]

[tool.poetry.dependencies]
python = "^3.11"
httpx = "^0.27.0"
pydantic = "^2.6.0"
pysolr = "^3.9.0"
tenacity = "^8.2.3"
fastapi = "^0.128.1"
uvicorn = "^0.40.0"
python-dotenv = "^1.2.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
black = "^24.0.0"
isort = "^5.13.0"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
